Revert "Merge pull request #34 from swcamp9thTeam1/feature/save-map"
This reverts commit d00f3fc439b4215c9f81272e503b19aa82f4ecf6, reversing
changes made to 7a61fed3af0bbac23fe29caaca91a610982646b7.
