+++ src/components/SearchBar.vue
