+++ src/data/good-stores.json
@@ -52,253 +52,7 @@
             "good": {
                 "menu": "짜장면",
                 "price": 6500
-            },
-            "accountBooks": [
-                {
-                    "accbookCode": 1,
-                    "title": "학원점심(짜장면)",
-                    "category": "식비",
-                    "asset": "신한체크카드",
-                    "createdAt": "2024-09-01",
-                    "amount": -7000
-                },
-                {
-                    "accbookCode": 2,
-                    "title": "동화장 짬뽕밥",
-                    "category": "식비",
-                    "asset": "신한체크카드",
-                    "createdAt": "2024-09-11",
-                    "amount": -8000
-                },
-                {
-                    "accbookCode": 3,
-                    "title": "잡채밥",
-                    "category": "식비",
-                    "asset": "신한체크카드",
-                    "createdAt": "2024-09-12",
-                    "amount": -9000
-                }
-            ]
-        }
-    ],
-    "store-reviews": [
-        {
-            "id": "1",
-            "priceAvg": 10500,
-            "reviews": [
-                {
-                    "reviewCode": 1,
-                    "nickname": "짜장면이최고야",
-                    "peopleNum": 2,
-                    "totalPrice": 35000,
-                    "content": "짬뽕이랑 탕수육은 무조건 드세요!!",
-                    "images": []
-                },
-                {
-                    "reviewCode": 2,
-                    "nickname": "짜장면덕후",
-                    "peopleNum": 1,
-                    "totalPrice": 7000,
-                    "content": "맛있어요.. 재방문 의사 100%",
-                    "images": [
-                        "https://i.namu.wiki/i/j2AxLP9AtrcJebh4DVfGxowfXwI3a95dG_YZb_Ktczc6Ca7ACyd_NJL3YHQMw8SABGTQiJDwSpySOSSBLZVEZw.webp"
-                    ]
-                },
-                {
-                    "reviewCode": 3,
-                    "nickname": "짜장보단짬뽕",
-                    "peopleNum": 1,
-                    "totalPrice": 8000,
-                    "content": "짬뽕 정말 맛있네요",
-                    "images": [
-                        "https://i.namu.wiki/i/upNZ7cYsFsAfU0KcguO6OHMK68xC-Bj8EXxdCti61Jhjx10UCBgdK5bZCEx41-aAWcjWZ5JMKFUSaUGLC1tqWg.webp",
-                        "https://i.namu.wiki/i/upNZ7cYsFsAfU0KcguO6OHMK68xC-Bj8EXxdCti61Jhjx10UCBgdK5bZCEx41-aAWcjWZ5JMKFUSaUGLC1tqWg.webp",
-                        "https://i.namu.wiki/i/upNZ7cYsFsAfU0KcguO6OHMK68xC-Bj8EXxdCti61Jhjx10UCBgdK5bZCEx41-aAWcjWZ5JMKFUSaUGLC1tqWg.webp"
-                    ]
-                }
-            ]
-        }
-    ],
-    "result": [
-        {
-            "dailyDate": "2024-09-01",
-            "totalOAmount": 15000,
-            "totalIAmount": 20000,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-02",
-            "totalOAmount": 30000,
-            "totalIAmount": 0,
-            "evaluation": "Bad"
-        },
-        {
-            "dailyDate": "2024-09-03",
-            "totalOAmount": 0,
-            "totalIAmount": 500000,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-04",
-            "totalOAmount": 25000,
-            "totalIAmount": 0,
-            "evaluation": "Bad"
-        },
-        {
-            "dailyDate": "2024-09-05",
-            "totalOAmount": 0,
-            "totalIAmount": 15000,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-06",
-            "totalOAmount": 2000,
-            "totalIAmount": 0,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-07",
-            "totalOAmount": 18000,
-            "totalIAmount": 100000,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-08",
-            "totalOAmount": 0,
-            "totalIAmount": 64000,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-09",
-            "totalOAmount": 77000,
-            "totalIAmount": 0,
-            "evaluation": "Bad"
-        },
-        {
-            "dailyDate": "2024-09-10",
-            "totalOAmount": 0,
-            "totalIAmount": 15000,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-11",
-            "totalOAmount": 25000,
-            "totalIAmount": 0,
-            "evaluation": "Bad"
-        },
-        {
-            "dailyDate": "2024-09-12",
-            "totalOAmount": 0,
-            "totalIAmount": 100000,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-13",
-            "totalOAmount": 30000,
-            "totalIAmount": 0,
-            "evaluation": "Bad"
-        },
-        {
-            "dailyDate": "2024-09-14",
-            "totalOAmount": 0,
-            "totalIAmount": 10000,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-15",
-            "totalOAmount": 8000,
-            "totalIAmount": 0,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-16",
-            "totalOAmount": 12000,
-            "totalIAmount": 12000,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-17",
-            "totalOAmount": 15000,
-            "totalIAmount": 0,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-18",
-            "totalOAmount": 0,
-            "totalIAmount": 9000,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-19",
-            "totalOAmount": 25000,
-            "totalIAmount": 0,
-            "evaluation": "Bad"
-        },
-        {
-            "dailyDate": "2024-09-20",
-            "totalOAmount": 0,
-            "totalIAmount": 300000,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-21",
-            "totalOAmount": 2000,
-            "totalIAmount": 0,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-22",
-            "totalOAmount": 0,
-            "totalIAmount": 15000,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-23",
-            "totalOAmount": 1000,
-            "totalIAmount": 0,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-24",
-            "totalOAmount": 10000,
-            "totalIAmount": 0,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-25",
-            "totalOAmount": 10000,
-            "totalIAmount": 0,
-            "evaluation": "Good"
-        },
-        {
-            "dailyDate": "2024-09-26",
-            "totalOAmount": 25000,
-            "totalIAmount": 0,
-            "evaluation": "Bad"
-        },
-        {
-            "dailyDate": "2024-09-27",
-            "totalOAmount": 0,
-            "totalIAmount": 0,
-            "evaluation": null
-        },
-        {
-            "dailyDate": "2024-09-28",
-            "totalOAmount": 0,
-            "totalIAmount": 0,
-            "evaluation": null
-        },
-        {
-            "dailyDate": "2024-09-29",
-            "totalOAmount": 0,
-            "totalIAmount": 0,
-            "evaluation": null
-        },
-        {
-            "dailyDate": "2024-09-30",
-            "totalOAmount": 0,
-            "totalIAmount": 0,
-            "evaluation": null
+            }
         }
     ]
 }
