+++ src/data/mockup.json
@@ -1,125 +1,46 @@
 {
-    "map-stores": [
+    "good-stores": [
         {
-            "id": 1,
+            "id": "1",
             "latitude": "37.499590490909185",
             "longitude": "127.0263723554437",
-            "isGood": true,
-            "goodPrice": 8500,
-            "visited": false,
-            "visitCnt": null,
-            "isExistAvg": false,
-            "onePersonCostAvg": null
+            "price": 8500
         },
         {
-            "id": 2,
+            "id": "2",
             "latitude": "37.499427948430814",
             "longitude": "127.02794423197847",
-            "isGood": true,
-            "goodPrice": 8300,
-            "visited": false,
-            "visitCnt": null,
-            "isExistAvg": true,
-            "onePersonCostAvg": 20000
+            "price": 8300
         },
         {
-            "id": 3,
+            "id": "3",
             "latitude": "37.498553760499505",
             "longitude": "127.02882598822454",
-            "isGood": true,
-            "goodPrice": 9500,
-            "visited": false,
-            "visitCnt": null,
-            "isExistAvg": true,
-            "onePersonCostAvg": 10000
+            "price": 9500
         },
         {
-            "id": 4,
+            "id": "4",
             "latitude": "37.497625593121384",
             "longitude": "127.02935713582038",
-            "isGood": true,
-            "goodPrice": 7500,
-            "visited": false,
-            "visitCnt": null,
-            "isExistAvg": true,
-            "onePersonCostAvg": 3400
+            "price": 7500
         },
         {
-            "id": 5,
+            "id": "5",
             "latitude": "37.49646391248451",
             "longitude": "127.02675574250912",
-            "isGood": true,
-            "goodPrice": 6500,
-            "visited": false,
-            "visitCnt": null,
-            "isExistAvg": false,
-            "onePersonCostAvg": null
+            "price": 6500
         },
         {
-            "id": 6,
+            "id": "6",
             "latitude": "37.49646391248451",
             "longitude": "127.02587362608637",
-            "isGood": true,
-            "goodPrice": 3500,
-            "visited": false,
-            "visitCnt": null,
-            "isExistAvg": true,
-            "onePersonCostAvg": 17000
+            "price": 3500
         },
         {
-            "id": 7,
+            "id": "7",
             "latitude": "37.49754540521486",
             "longitude": "127.02546694890695",
-            "isGood": true,
-            "goodPrice": 7900,
-            "visited": false,
-            "visitCnt": null,
-            "isExistAvg": true,
-            "onePersonCostAvg": 13400
-        },
-        {
-            "id": 8,
-            "latitude": "37.497535461505684",
-            "longitude": "127.02948149502778",
-            "isGood": false,
-            "goodPrice": null,
-            "visited": true,
-            "visitCnt": 1,
-            "isExistAvg": false,
-            "onePersonCostAvg": null
-        },
-        {
-            "id": 9,
-            "latitude": "37.49671536281186",
-            "longitude": "127.03020491448352",
-            "isGood": false,
-            "goodPrice": null,
-            "visited": true,
-            "visitCnt": 2,
-            "isExistAvg": false,
-            "onePersonCostAvg": null
-        },
-        {
-            "id": 10,
-            "latitude": "37.496201943633714",
-            "longitude": "127.02959405469642",
-            "isGood": false,
-            "goodPrice": null,
-            "visited": true,
-            "visitCnt": 3,
-            "isExistAvg": true,
-            "onePersonCostAvg": 35000
-        },
-        {
-            "id": 11,
-            "latitude": "37.49640072567703",
-            "longitude": "127.02726459882308",
-            "isGood": false,
-            "goodPrice": null,
-            "visited": true,
-            "visitCnt": 1,
-            "isExistAvg": false,
-            "onePersonCostAvg": null
+            "price": 7900
         }
     ],
 
@@ -133,45 +54,6 @@
                 "menu": "짜장면",
                 "price": 6500
             }
-        }
-    ],
-
-    "store-reviews": [
-        {
-            "id": "1",
-            "priceAvg": 10500,
-            "reviews": [
-                {
-                    "id": 1,
-                    "nickname": "짜장면이최고야",
-                    "peopleNum": 2,
-                    "totalPrice": 35000,
-                    "content": "짬뽕이랑 탕수육은 무조건 드세요!!",
-                    "images": []
-                },
-                {
-                    "id": 2,
-                    "nickname": "짜장면덕후",
-                    "peopleNum": 1,
-                    "totalPrice": 7000,
-                    "content": "맛있어요.. 재방문 의사 100%",
-                    "images": [
-                        "https://i.namu.wiki/i/j2AxLP9AtrcJebh4DVfGxowfXwI3a95dG_YZb_Ktczc6Ca7ACyd_NJL3YHQMw8SABGTQiJDwSpySOSSBLZVEZw.webp"
-                    ]
-                },
-                {
-                    "id": 3,
-                    "nickname": "짜장보단짬뽕",
-                    "peopleNum": 1,
-                    "totalPrice": 8000,
-                    "content": "짬뽕 정말 맛있네요",
-                    "images": [
-                        "https://i.namu.wiki/i/upNZ7cYsFsAfU0KcguO6OHMK68xC-Bj8EXxdCti61Jhjx10UCBgdK5bZCEx41-aAWcjWZ5JMKFUSaUGLC1tqWg.webp",
-                        "https://cdn.pixabay.com/photo/2021/02/11/01/58/jambong-6004032_1280.jpg",
-                        "https://i.namu.wiki/i/upNZ7cYsFsAfU0KcguO6OHMK68xC-Bj8EXxdCti61Jhjx10UCBgdK5bZCEx41-aAWcjWZ5JMKFUSaUGLC1tqWg.webp"
-                    ]
-                }
-            ]
         }
     ],
